+++ src/node/node_path.rs
@@ -1,13 +1,14 @@
 use crate::storage::StorePrefix;
 use derive_more::Display;
+use std::path::PathBuf;
 use thiserror::Error;
 
 /// A Zarr hierarchy node path.
 ///
 /// See <https://zarr-specs.readthedocs.io/en/latest/v3/core/v3.0.html#path>
 #[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Display)]
-#[display(fmt = "{_0}")]
-pub struct NodePath(String);
+#[display(fmt = "{_0:?}")]
+pub struct NodePath(PathBuf);
 
 /// An invalid node path.
 #[derive(Debug, Error)]
@@ -22,7 +23,7 @@ impl NodePath {
     /// Returns [`NodePathError`] if `path` is not valid according to [`NodePath::validate`()].
     pub fn new(path: &str) -> Result<NodePath, NodePathError> {
         if NodePath::validate(path) {
-            Ok(NodePath(path.to_string()))
+            Ok(NodePath(PathBuf::from(path)))
         } else {
             Err(NodePathError(path.to_string()))
         }
@@ -31,12 +32,19 @@ impl NodePath {
     /// The root node.
     #[must_use]
     pub fn root() -> NodePath {
-        NodePath("/".to_string())
+        NodePath(PathBuf::from("/"))
     }
 
     /// Extracts a string slice containing the node path `String`.
+    #[allow(clippy::missing_panics_doc)]
     #[must_use]
     pub fn as_str(&self) -> &str {
+        self.0.to_str().unwrap()
+    }
+
+    /// Extracts the path as a [`std::path::Path`].
+    #[must_use]
+    pub fn as_path(&self) -> &std::path::Path {
         &self.0
     }